+++ ICD.Connect.Audio.Shure/ShureMxaRouteSourceControl.cs
@@ -1,0 +1,57 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using ICD.Common.Utils.Services;
+using ICD.Connect.Routing;
+using ICD.Connect.Routing.Connections;
+using ICD.Connect.Routing.Controls;
+using ICD.Connect.Routing.EventArguments;
+using ICD.Connect.Routing.RoutingGraphs;
+
+namespace ICD.Connect.Audio.Shure
+{
+	public sealed class ShureMxaRouteSourceControl : AbstractRouteSourceControl<IShureMxaDevice>
+	{
+		public override event EventHandler<TransmissionStateEventArgs> OnActiveTransmissionStateChanged;
+
+		/// <summary>
+		/// Constructor.
+		/// </summary>
+		/// <param name="parent"></param>
+		/// <param name="id"></param>
+		public ShureMxaRouteSourceControl(IShureMxaDevice parent, int id)
+			: base(parent, id)
+		{
+		}
+
+		#region Methods
+
+		/// <summary>
+		/// Returns true if the device is actively transmitting on the given output.
+		/// This is NOT the same as sending video, since some devices may send an
+		/// idle signal by default.
+		/// </summary>
+		/// <param name="output"></param>
+		/// <param name="type"></param>
+		/// <returns></returns>
+		public override bool GetActiveTransmissionState(int output, eConnectionType type)
+		{
+			return true;
+		}
+
+		/// <summary>
+		/// Returns the outputs.
+		/// </summary>
+		/// <returns></returns>
+		public override IEnumerable<ConnectorInfo> GetOutputs()
+		{
+			return
+				ServiceProvider.GetService<IRoutingGraph>()
+				               .Connections.GetChildren()
+				               .Where(c => c.Source.Device == Parent.Id && c.Source.Control == Id)
+				               .Select(c => new ConnectorInfo(c.Source.Address, eConnectionType.Audio));
+		}
+
+		#endregion
+	}
+}